+++ include/backend/x11.h
@@ -6,9 +6,9 @@
 #include <X11/Xlib-xcb.h>
 #include <wayland-server.h>
 #include <wlr/render/egl.h>
+#include <wlr/render/render.h>
 #include <wlr/types/wlr_output.h>
 #include <wlr/types/wlr_input_device.h>
-#include "render/render.h"
 
 struct wlr_x11_backend;
