+++ include/rootston/server.h
@@ -3,8 +3,8 @@
 #include <wayland-server.h>
 #include <wlr/backend.h>
 #include <wlr/backend/session.h>
+#include <wlr/render/render.h>
 #include <wlr/types/wlr_data_device.h>
-#include "render/render.h"
 #ifdef HAS_XWAYLAND
 #include <wlr/xwayland.h>
 #endif
