+++ include/wlr/types/wlr_compositor.h
@@ -2,7 +2,7 @@
 #define WLR_TYPES_WLR_COMPOSITOR_H
 
 #include <wayland-server.h>
-#include "render/render.h"
+#include <wlr/render/render.h>
 
 struct wlr_compositor {
 	struct wl_global *wl_global;
