+++ include/wlr/types/wlr_surface.h
@@ -5,7 +5,7 @@
 #include <pixman.h>
 #include <stdint.h>
 #include <stdbool.h>
-#include "render/render.h"
+#include <wlr/render/render.h>
 
 struct wlr_frame_callback {
 	struct wl_resource *resource;
